+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Changes to procedures and constraints:
     * Units for width and height are moved next to the width and height widgets, respectively.
     * More units beside pixels and percentages can be chosen from (the same ones available in GIMP). These units are recalculated automatically.
     * Switching between pixels, percentages and other units will now restore the last value for that unit.
+    * Merged `Scale to fit`, `Keep aspect ratio` and `Dimension to keep` to a single widget.
     * Added a new option to change the image resolution.
 * Edit Layers: The `Selected in GIMP` constraint is now added on the first run/after reset.